v1.0 is ready, preparation for merge with dev
